[doc] Instructions to set up a slave
Send email for a password, and point to the official buildbot docs on
creating a buildslave for the latest version.

The specific links for Unix based and Windows have short and clear
instructions, but the manual is useful if you aren't used to the command
line (e.g., you need to read about the parameters of buildslave
create-slave) or if you want to tweak the configuration.

- Related: #7354